feat(mcp): add token-gated Streamable HTTP transport for URL-only clients
Extract the tool definitions and handlers from mcp/server.ts into a
shared factory (mcp/library-server.ts) and serve them over a second
transport at POST /mcp on the app server. Each request gets a stateless
server instance whose handlers forward to the same /api/library/*
routes the stdio shim uses, so both transports share one execution
path.

The endpoint stays on the loopback bind and requires a bearer token
generated at boot into ~/.stashbase/mcp-http-token (0600): a spawned
stdio server is reachable only by whoever spawned it, but a loopback
URL can be probed by any local process, so possession of the token file
marks a caller as the local user. GET and DELETE answer 405 - the tool
surface is pure request/response, so there are no sessions or
server-initiated streams to resume. Same-machine Docker clients attach
via http://host.docker.internal:8090/mcp.

Closes #17.
